Move Grape::Endpoint#declared specs into own spec file
